backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
python-multipart
pypdf2
python-docx